// src/Custom/pages/Companies/Company/Form/interfaces.ts
export enum CompanyFormSteps {
  // Вводная информация для пользователя, чтобы он знал что и как тут можно делать
  Intro = 'intro',
  // Пользователь наполняет черновой вариант, или редактирует уже существующу компанию
  ClientFill = 'fill',
  // Пользователь редактирует фотогалерею заведения
  ClientEditGallery = 'gallery',
  // Ревью. Тут клиент видит свою карточку и сообщение, что будет опубликована после проверки.
  // Тут еще можно попросить предоставить контактные данные, чтобы мы могли связаться и сообщить о статусе публикации
  Preview = 'preview',
  // Карточка на ревью
  Review = 'review',
}